fix extended statistics from commit c5ce81dc48
Some of the MySQL queries will return NULL values, if there
are no downloads yet. Passing these values to the template
system is possible, but the values will not get replaced
properly in the template. This commit fixes that by
replacing NULL results with zeroes (0) before passing them
to the template system.